Add basic support for DNSSEC RR types to inet_res
This adds support for querying for and receiving DS, DNSKEY, RRSIG,
NSEC, NSEC3 and NSEC3PARAM RR types in DNS queries. The aim is to
provide as basis that can be used to for example implement a validating
recursive resolver as a third-party module on top of this, without
having to re-implement support for basic DNS. All cryptographic material
here is left as binaries, so they can be appropriately interpreted
according to indicated algorithms.